fix(consensus): always send finalized tip to execution layer (#618)
Always sends the latest finalized tip that the consensus engine knows
about to the execution layer. This is to ensure that the execution layer
can trigger an e2e backfill and does not rely on the consensus engine
trickle feeding blocks to it.

Makes use a the new feature introduced in
https://github.com/commonwarexyz/monorepo/pull/2065.

# Follow-up

With this patch we would still trickle feed individual blocks from the
consensus engine to the execution layer. A followup will skip this as to
not unnecessarily calculate state roots.

---------

Co-authored-by: joshieDo <93316087+joshieDo@users.noreply.github.com>