fix(consensus): DKGs with failed shares are still successful, on-chain DKG is source of truth (#1512)
This patch ensures that the full DKG ceremony isn't failed if a player
can't get its shares from on-chain data.

The reasoning is simple: if the overall DKG ceremony was a success, it
means that a quorum of dealers and players have successfully
participated. The case of DKG success && player failure (for all
players) is not possible unless all critical number of nodes crashed and
had critical data loss at the same time.

Furthermore, this patch handles the case of a DKG outcome mismatch
between what was published to chain and what the network came to
consensus over, and what a node decides locally: if there is a mismatch,
the node will always select what was published on chain.